Add functionality to the TrajectoryManager

+++ SymplePlanner/src/main/java/top/symple/sympleplanner/interfaces/IDriveTrain.kt
@@ -9,5 +9,5 @@ interface IDriveTrain {
      *
      * @param power Array of motor powers.
      */
-    fun setPower(power: IntArray);
+    fun setPower(power: DoubleArray);
 }
